Added Responsiveness and reconstructed navbar
