+++ package.json
@@ -19,15 +19,17 @@
     "@fortawesome/free-solid-svg-icons": "^5.13.1",
     "@fortawesome/react-fontawesome": "^0.1.11",
     "@tippyjs/react": "^4.1.0",
+    "bootstrap": "^4.5.0",
     "gatsby": "^2.23.12",
     "gatsby-image": "^2.4.12",
     "gatsby-plugin-fontawesome-css": "^1.0.0",
     "gatsby-plugin-sharp": "^2.6.18",
     "gatsby-source-filesystem": "^2.3.18",
     "gatsby-transformer-sharp": "^2.5.10",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-scrollchor": "^6.0.0"
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-scrollchor": "^6.0.0",
+    "reactstrap": "^8.5.1"
   },
   "devDependencies": {
     "prettier": "2.0.5"
