--- src/components/NavBar.css
@@ -1,31 +1,0 @@
-nav { 
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    background-color: rgb(219, 171, 127);
-    height: 50px;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width:100%;
-    padding: 0;
-    z-index: 100;
-    box-shadow: 0px 4px 8px 0px rgba(0,0,0,0.2);
-
-}
-.nav-item {
-    margin: 15px;
-    display: block;
-    color: white;
-    font-size: 17px;
-    text-decoration: none;
-}
-
-.nav-item:hover {
-    color: gray;
-}
-
-.logo {
-    margin-top: -12px;
-    width: 45px;
-}
