+++ src/components/Projects.css
@@ -14,3 +14,15 @@
 .projects h1 {
     text-decoration: underline;
 }
+
+@media all and (max-width: 1024px) {
+    .projectContainer {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+}
+
+
+ 
+
