+++ src/components/Technologies.css
@@ -2,6 +2,7 @@
     display: flex;
     flex-direction: row;
     justify-content: center;
+    max-width: 1em;
 }
 
 .heading {
@@ -9,3 +10,10 @@
     text-decoration: underline;
     margin-top: 100px;
 }
+
+@media all and (max-width: 768px) {
+   .container {
+    flex-wrap: wrap;
+   }
+}
+
