+++ src/pages/index.css
@@ -1,0 +1,4 @@
+@import url("https://fonts.googleapis.com/css?family=Source+Sans+Pro");
+body {
+ font-family: "Source Sans Pro"
+}
